pyproject.toml
[project]
name = "betting-odds-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Rui Pereira", email = "ruiarpereira15@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "selenium>=4.41.0",
    "tzdata>=2025.3",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
